+++ src/util.rs
@@ -1,4 +1,3 @@
-use core::ffi::c_int;
 use core::ptr::{addr_of, addr_of_mut, null_mut};
 
 use std::sync::atomic::{AtomicU32, Ordering};
@@ -71,8 +70,8 @@ const ISZ_PREF_ALREADY_PLAYED: usize = 17;
 
 const CCH_MSG_MAX: usize = 128;
 const CCH_MAX_PATHNAME: usize = 250;
-const F_RESIZE: c_int = 0x02;
-const FMENU_FLAG_OFF: c_int = 0x01;
+const F_RESIZE: i32 = 0x02;
+const FMENU_FLAG_OFF: i32 = 0x01;
 
 const SZ_INI_FILE: PCWSTR = w!("entpack.ini");
 
@@ -81,14 +80,14 @@ fn seed_rng(seed: u32) {
     RNG_STATE.store(value, Ordering::Relaxed);
 }
 
-fn next_rand() -> c_int {
+fn next_rand() -> i32 {
     let mut current = RNG_STATE.load(Ordering::Relaxed);
     loop {
         let next = current
             .wrapping_mul(RNG_MULTIPLIER)
             .wrapping_add(RNG_INCREMENT);
         match RNG_STATE.compare_exchange(current, next, Ordering::Relaxed, Ordering::Relaxed) {
-            Ok(_) => return ((next >> 16) & 0x7FFF) as c_int,
+            Ok(_) => return ((next >> 16) & 0x7FFF) as i32,
             Err(actual) => current = actual,
         }
     }
@@ -114,7 +113,7 @@ unsafe fn default_name_ptr() -> PCWSTR {
     addr_of!(szDefaultName[0])
 }
 
-fn clamp(value: c_int, min: c_int, max: c_int) -> c_int {
+fn clamp(value: i32, min: i32, max: i32) -> i32 {
     value.max(min).min(max)
 }
 
@@ -122,7 +121,7 @@ fn make_int_resource(id: u16) -> PCWSTR {
     id as usize as *const u16
 }
 
-pub fn Rnd(rnd_max: c_int) -> c_int {
+pub fn Rnd(rnd_max: i32) -> i32 {
     // Return a pseudo-random number in the [0, rnd_max) range like the C helper did.
     if rnd_max <= 0 {
         0
@@ -150,7 +149,7 @@ pub unsafe fn ReportErr(id_err: u16) {
             sz_title.as_mut_ptr(),
             CCH_MSG_MAX as i32,
         );
-        wsprintfW(sz_msg.as_mut_ptr(), sz_title.as_ptr(), id_err as c_int);
+        wsprintfW(sz_msg.as_mut_ptr(), sz_title.as_ptr(), id_err as i32);
     }
 
     LoadStringW(
@@ -175,11 +174,11 @@ pub unsafe fn LoadSz(id: u16, sz: *mut u16, cch: u32) {
 }
 
 pub unsafe fn ReadIniInt(
-    isz_pref: c_int,
-    val_default: c_int,
-    val_min: c_int,
-    val_max: c_int,
-) -> c_int {
+    isz_pref: i32,
+    val_default: i32,
+    val_min: i32,
+    val_max: i32,
+) -> i32 {
     // Pull an integer from the legacy .ini file, honoring the same clamp the game always used.
     if isz_pref < 0 || (isz_pref as usize) >= ISZ_PREF_MAX {
         return val_default;
@@ -190,11 +189,11 @@ pub unsafe fn ReadIniInt(
         return val_default;
     }
 
-    let value = GetPrivateProfileIntW(class_ptr(), key, val_default, SZ_INI_FILE) as c_int;
+    let value = GetPrivateProfileIntW(class_ptr(), key, val_default, SZ_INI_FILE) as i32;
     clamp(value, val_min, val_max)
 }
 
-pub unsafe fn ReadIniSz(isz_pref: c_int, sz_ret: *mut u16) {
+pub unsafe fn ReadIniSz(isz_pref: i32, sz_ret: *mut u16) {
     // Grab the string from entpack.ini or fall back to the default Hall of Fame name.
     if sz_ret.is_null() || isz_pref < 0 || (isz_pref as usize) >= ISZ_PREF_MAX {
         return;
@@ -257,7 +256,7 @@ pub unsafe fn InitConst() {
     ) == 0
     {
         g_hReg = key;
-        already_played = ReadInt(ISZ_PREF_ALREADY_PLAYED as c_int, 0, 0, 1);
+        already_played = ReadInt(ISZ_PREF_ALREADY_PLAYED as i32, 0, 0, 1);
         RegCloseKey(key);
         g_hReg = std::ptr::null_mut();
     }
@@ -268,36 +267,36 @@ pub unsafe fn InitConst() {
 
     let prefs = prefs_mut();
 
-    (*prefs).Height = ReadIniInt(ISZ_PREF_HEIGHT as c_int, MINHEIGHT, DEFHEIGHT, 25);
-    (*prefs).Width = ReadIniInt(ISZ_PREF_WIDTH as c_int, MINWIDTH, DEFWIDTH, 30);
+    (*prefs).Height = ReadIniInt(ISZ_PREF_HEIGHT as i32, MINHEIGHT, DEFHEIGHT, 25);
+    (*prefs).Width = ReadIniInt(ISZ_PREF_WIDTH as i32, MINWIDTH, DEFWIDTH, 30);
     (*prefs).wGameType = ReadIniInt(
-        ISZ_PREF_GAME as c_int,
+        ISZ_PREF_GAME as i32,
         WGAME_BEGIN,
         WGAME_BEGIN,
         WGAME_EXPERT + 1,
     ) as u16;
-    (*prefs).Mines = ReadIniInt(ISZ_PREF_MINES as c_int, 10, 10, 999);
-    (*prefs).xWindow = ReadIniInt(ISZ_PREF_XWINDOW as c_int, 80, 0, 1024);
-    (*prefs).yWindow = ReadIniInt(ISZ_PREF_YWINDOW as c_int, 80, 0, 1024);
+    (*prefs).Mines = ReadIniInt(ISZ_PREF_MINES as i32, 10, 10, 999);
+    (*prefs).xWindow = ReadIniInt(ISZ_PREF_XWINDOW as i32, 80, 0, 1024);
+    (*prefs).yWindow = ReadIniInt(ISZ_PREF_YWINDOW as i32, 80, 0, 1024);
 
-    (*prefs).fSound = ReadIniInt(ISZ_PREF_SOUND as c_int, 0, 0, FSOUND_ON);
-    (*prefs).fMark = bool_from_int(ReadIniInt(ISZ_PREF_MARK as c_int, TRUE as c_int, 0, 1));
-    (*prefs).fTick = bool_from_int(ReadIniInt(ISZ_PREF_TICK as c_int, 0, 0, 1));
+    (*prefs).fSound = ReadIniInt(ISZ_PREF_SOUND as i32, 0, 0, FSOUND_ON);
+    (*prefs).fMark = bool_from_int(ReadIniInt(ISZ_PREF_MARK as i32, TRUE, 0, 1));
+    (*prefs).fTick = bool_from_int(ReadIniInt(ISZ_PREF_TICK as i32, 0, 0, 1));
     (*prefs).fMenu = ReadIniInt(
-        ISZ_PREF_MENU as c_int,
+        ISZ_PREF_MENU as i32,
         FMENU_ALWAYS_ON,
         FMENU_ALWAYS_ON,
         FMENU_ON,
     );
 
-    (*prefs).rgTime[WGAME_BEGIN as usize] = ReadIniInt(ISZ_PREF_BEGIN_TIME as c_int, 999, 0, 999);
-    (*prefs).rgTime[WGAME_INTER as usize] = ReadIniInt(ISZ_PREF_INTER_TIME as c_int, 999, 0, 999);
-    (*prefs).rgTime[WGAME_EXPERT as usize] = ReadIniInt(ISZ_PREF_EXPERT_TIME as c_int, 999, 0, 999);
+    (*prefs).rgTime[WGAME_BEGIN as usize] = ReadIniInt(ISZ_PREF_BEGIN_TIME as i32, 999, 0, 999);
+    (*prefs).rgTime[WGAME_INTER as usize] = ReadIniInt(ISZ_PREF_INTER_TIME as i32, 999, 0, 999);
+    (*prefs).rgTime[WGAME_EXPERT as usize] = ReadIniInt(ISZ_PREF_EXPERT_TIME as i32, 999, 0, 999);
 
-    ReadIniSz(ISZ_PREF_BEGIN_NAME as c_int, (*prefs).szBegin.as_mut_ptr());
-    ReadIniSz(ISZ_PREF_INTER_NAME as c_int, (*prefs).szInter.as_mut_ptr());
+    ReadIniSz(ISZ_PREF_BEGIN_NAME as i32, (*prefs).szBegin.as_mut_ptr());
+    ReadIniSz(ISZ_PREF_INTER_NAME as i32, (*prefs).szInter.as_mut_ptr());
     ReadIniSz(
-        ISZ_PREF_EXPERT_NAME as c_int,
+        ISZ_PREF_EXPERT_NAME as i32,
         (*prefs).szExpert.as_mut_ptr(),
     );
 
@@ -312,8 +311,8 @@ pub unsafe fn InitConst() {
         ReleaseDC(desktop, hdc);
     }
     (*prefs).fColor = bool_from_int(ReadIniInt(
-        ISZ_PREF_COLOR as c_int,
-        default_color as c_int,
+        ISZ_PREF_COLOR as i32,
+        default_color as i32,
         0,
         1,
     ));
@@ -338,7 +337,7 @@ pub unsafe fn CheckEm(idm: u16, f_check: BOOL) {
     );
 }
 
-pub unsafe fn SetMenuBar(f_active: c_int) {
+pub unsafe fn SetMenuBar(f_active: i32) {
     // Persist the menu visibility preference, refresh accelerator state, and resize the window.
     (*prefs_mut()).fMenu = f_active;
     FixMenus();
@@ -403,15 +402,15 @@ pub unsafe fn DoHelp(w_command: u16, l_param: u32) {
     HtmlHelpA(GetDesktopWindow(), buffer.as_ptr() as PCSTR, l_param, 0);
 }
 
-pub unsafe fn GetDlgInt(h_dlg: HWND, dlg_id: c_int, num_lo: c_int, num_hi: c_int) -> c_int {
+pub unsafe fn GetDlgInt(h_dlg: HWND, dlg_id: i32, num_lo: i32, num_hi: i32) -> i32 {
     // Mirror GetDlgInt from util.c: clamp user input to the legal range before the caller consumes it.
     let mut success = 0i32;
     let value = GetDlgItemInt(h_dlg, dlg_id, &mut success, FALSE);
-    let value = value as c_int;
+    let value = value as i32;
     clamp(value, num_lo, num_hi)
 }
 
-fn bool_from_int(value: c_int) -> BOOL {
+fn bool_from_int(value: i32) -> BOOL {
     if value != 0 {
         TRUE
     } else {
